fastapi-app/requirements.txt
fastapi
uvicorn[standard]
httpx
python-dotenv
pydantic
llama-cpp-python
sentence-transformers
psycopg2-binary
